% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{assessment_part1}
\alias{assessment_part1}
\title{Assessment object resulting from harsat procedure}
\format{
An object of class \code{list} of length 6.
}
\usage{
assessment_part1
}
\description{
This object contains all the results from an assessment, including both raw data, metadata as well as effects of trend analysis. See \code{?run_assessment_tar} for
the full code for making this object.
}
\keyword{datasets}
