% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_trend_text.R
\name{get_trend_text}
\alias{get_trend_text}
\title{Extract trend text from assessment data object}
\usage{
get_trend_text(assessment_data, trendlength)
}
\arguments{
\item{assessment_data}{One list item (i.e. one time series) from a assessment data object, i.e. the output of
either 'get_assessment_data' or 'combine_assessment_data'.}

\item{trendlength}{Must be either 'overall' or 'recent'}
}
\value{
A string (i.e., a text)
}
\description{
Extract trend text from assessment data object
}
\examples{
#' #' # Extract data for plotting from an assessment object
assessment_data <- get_assessment_data(assessment_part1)
# Get trend text
get_trend_text(assessment_data[["4994 CD Gadus morhua LI NA"]], "overall")
get_trend_text(assessment_data[["4994 CD Gadus morhua LI NA"]], "recent")
}
