% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_assessment.R
\name{ggplot_assessment}
\alias{ggplot_assessment}
\title{Plot assessment graph using ggplot2}
\usage{
ggplot_assessment(
  assessment_data,
  plot_points = "annual",
  logscale = TRUE,
  pointcolor = "darkred",
  pointshapes = c(19, 6),
  trendcolor_line = "darkblue",
  trendcolor_fill = "lightblue",
  trendwidth = 0.8,
  ylim = NULL,
  title = NULL,
  subtitle = NULL,
  add_trend_text = FALSE
)
}
\arguments{
\item{assessment_data}{One list item (i.e. one time series) from a assessment data object, i.e. the output of
either 'get_assessment_data' or 'combine_assessment_data'.}

\item{plot_points}{Either "annual" or "all"}

\item{logscale}{Boolean (TRUE/FALSE); the default is TRUE}

\item{pointcolor}{String (color name or hex code) giving the color of the observation points}

\item{pointshapes}{A vector of two integers, giving the shape of the observation points for over-LOQ data (number 1)
and under-LOQ data (number 2). See ?points for the shapes corresponding to each number. The default is 19 (round dot)
for over-LOQ concentrations and 6 (triangle pointing down) for under-LOQ concentrations}

\item{trendcolor_line}{String (color name or hex code) giving the color of the trend line}

\item{trendcolor_fill}{String (color name or hex code) giving the fill color of the trend confidence interval}

\item{trendwidth}{Number giving the line with of the trend line}

\item{ylim}{A vector of two numbers, giving min and max of the y axis. If NULL (the default), this is determined automatically}

\item{title}{Plot title (by default, \code{output_id})}

\item{subtitle}{Plot subtitle (by default, \code{station_name})}
}
\value{
A ggplot2 object
}
\description{
Plot assessment graph using ggplot2
}
\details{
Makes a time trend plot, showing the estimated trend line, as well as points showing either all
concentrations (mulitiple samples per year) or the annual mean concentration.
}
\examples{

# Extract data for plotting from an assessment object
assessment_data <- get_assessment_data(assessment_part1)

# Check names of time series
names(assessment_part1$assessment)

# Plot one time series, showing trend and annual means, and logarithmic y axis
ggplot_assessment(assessment_data[["4994 CD Gadus morhua LI NA"]])

# Showing trend only
ggplot_assessment(assessment_data[["4994 CD Gadus morhua LI NA"]], plot_points = NA)

# Showing trend and raw data
ggplot_assessment(assessment_data[["4994 CD Gadus morhua LI NA"]], plot_points = "all")

# Same, but using linear logarithmic y axis
ggplot_assessment(assessment_data[["4994 CD Gadus morhua LI NA"]], plot_points = "all", logscale = FALSE)

# Add custom title and subtitle
ggplot_assessment(assessment_data[["4994 CD Gadus morhua LI NA"]], plot_points = "all",
                  title = 'Cadmium in cod liver at Bjørneroya station',
                  subtitle = 'Note: logarithmic y axis')

# Combine assessment data (from the branches of the targets work flow) and plot
\dontrun{
assessment_data <- combine_assessment_data()
ggplot_assessment(assessment_data[["10921 CD Gadus morhua LI NA"]])

# Plot all concentration data
ggplot_assessment(assessment_data[["10921 CD Gadus morhua LI NA"]], plot_points = "all")

# Plot all concentration data, use linear y scale (not log)
ggplot_assessment(assessment_data[["10921 CD Gadus morhua LI NA"]], plot_points = "all", logscale = FALSE)
}

}
