% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harsatextras_example.R
\name{harsatextras_example}
\alias{harsatextras_example}
\title{Convenience function to make example files easy to access}
\usage{
harsatextras_example(path = NULL)
}
\arguments{
\item{path}{A string. If \code{path} is not given, the function will return a list of the raw csv files in the package. If the name of a file is given as \code{path}, it will return the full path of the file.}
}
\value{
A list of file names of path is not given a string with the full path if path is given.
}
\description{
Convenience function to make example files easy to access
}
\examples{

harsatextras_example()
harsatextras_example("ICES_DOME_STATIONS_20230829_NO.csv")

# For reading a file
test <- read.csv(harsatextras_example("ICES_DOME_STATIONS_20230829_NO.csv"))
head(test, 2)

}
