% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{info_object}
\alias{info_object}
\title{Information object used by harsat procedure}
\format{
a list of 17 objects, including strings, lists and data frames
}
\usage{
info_object
}
\description{
This object contains a lot of information needed for assessment, including standard units, transformation to be used etc. for each determinand, standard dry weights
and lipid percentages for different tissues in each species, etc. This list is used in the tidying of the data as well as in
the assessment.
}
\keyword{datasets}
