% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{norway_branching_groups}
\alias{norway_branching_groups}
\title{Data frame defining groups for static branching}
\format{
A data frame with 7 columns, six columns that will be matched with the data ("determinand",
"station_code", "species", "matrix", "subseries", and "basis"), and one column that defines the groupings ("branch").
}
\usage{
norway_branching_groups
}
\description{
A small example data frame of four rows, containing data that defines how the four
time series of 'norway_timeseries_all' should be divided into groups. \code{split_timeseries_object} requires such a file
to define how the times series object should be split. In this case, the column 'branch' has two values
'CD_allstations' and 'PFOS_allstations', meaning that the time series object will be split into two groups, one with all CD (cadmium) tiem series
and one with all PFOS time series.
}
\keyword{datasets}
