% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{norway_timeseries_all}
\alias{norway_timeseries_all}
\title{Time series object resulting from harsat::create_timeseries()}
\format{
A list with six elements: \code{call}, \code{call.data},
\code{info}, \code{data}, \code{stations} and \code{timeSeries}.
}
\usage{
norway_timeseries_all
}
\description{
A small example time series object, containing data for four Norwegian
time series of contaminants in biota:
two series for cadmium (determinant = 'CD') and two for PFOS
(determinand = 'PFOS'). This object is the output of harsat::create_timeseries(),
and can be used as input for harsat::run_assessment. It can also be the input of
split_timeseries_object() in this package
}
\keyword{datasets}
