% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_tar.R
\name{read_data_tar}
\alias{read_data_tar}
\title{Modified version of harsat::read_data, for targets}
\usage{
read_data_tar(
  compartment = c("biota", "sediment", "water"),
  purpose = c("OSPAR", "HELCOM", "AMAP", "custom"),
  contaminants,
  data_format = c("ICES", "external"),
  filename_info,
  filename_stations,
  filename_contaminants,
  extraction = NULL,
  max_year = NULL,
  oddity_dir = "oddities",
  control = list()
)
}
\arguments{
\item{compartment}{See ?harsat::read_data}

\item{purpose}{See ?harsat::read_data}

\item{contaminants}{See ?harsat::read_data}

\item{data_format}{See ?harsat::read_data}

\item{filename_info}{Path to 'info' R object file (rds), see ?info_object}

\item{filename_stations}{Path to station dictionary file, csv format. See ?harsat::read_data for more}

\item{filename_contaminants}{Path to raw data file, csv format. See ?harsat::read_data for more}

\item{extraction}{See ?harsat::read_data}

\item{max_year}{See ?harsat::read_data}

\item{oddity_dir}{See ?harsat::read_data}

\item{control}{See ?harsat::read_data}
}
\value{
A list with four components: \code{call}, \code{info}, \code{data}, and \code{stations}
}
\description{
Modified version of harsat::read_data, for targets
}
\examples{

info_path <- harsatextras_example("info.rds")
stations_path <- harsatextras_example("ICES_DOME_STATIONS_20230829_NO.csv")
rawdata_path <- harsatextras_example("norway_rawdata.csv")
biota_data <- read_data_tar(
  compartment = "biota",
  purpose = "OSPAR",
  contaminants = file,
  data_format = "external",
  filename_info = info_path,
  filename_stations = stations_path,
  filename_contaminants = rawdata_path
  )

}
