% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_timeseries_object.R
\name{split_timeseries_object}
\alias{split_timeseries_object}
\title{Split a time series object into a list of time series objects}
\usage{
split_timeseries_object(object, df_branching, groupcolumn = "branch")
}
\arguments{
\item{object}{A time series object resulting from \code{harsat::create_timeseries()}. The object should be a list with 6 objects, named \code{call}, \code{call.data},
\code{info}, \code{data}, \code{stations} and \code{timeSeries}}

\item{df_branching}{A data frame containing (at least) 7 columns. The six columns that will be matched with the data are "determinand",
"station_code", "species", "matrix", "subseries", and "basis", and they must have exactly these names. The name of the last column, which will be used for grouping,
is given by \code{groupcolumn}; the default name is "branch"}

\item{groupcolumn}{A string with the name of the column that shall be used for splitting}
}
\value{
Returns a list of time series objects, each of which can be input for run_assessment_tar()
}
\description{
Split a time series object into a list of time series objects
}
\examples{

# Original time series object: a list of 6 elements
str(norway_timeseries_all, 1)

# Data frame defining groups (in column 'branch')
branching_path <- harsatextras_example("norway_branching_groups.csv")
norway_branching_groups <- read.csv(branching_path)

# Split time series object, using 'norway_branching_groups':
timeseries_list <- split_timeseries_object(norway_timeseries_all, norway_branching_groups)

# The result is a list of two time series objects, each bein a list of 6 elements:
str(timeseries_list, 2)

See ?run_assessment_tar for full workflow


}
