% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_data2.R
\name{tidy_data2}
\alias{tidy_data2}
\title{Tidy the data a bit more}
\usage{
tidy_data2(data)
}
\arguments{
\item{data}{The tidied data, typically returned from \code{tidy_data()}}
}
\value{
A harsat object similar to what is returned by tidy_data, just without the columns 'country' and 'station_name' in the 'data' object
}
\description{
Tidy the data a bit more
}
\examples{

\dontrun{
# Tidy data and create time series object
norway_tidy <- harsat::tidy_data(norway_data)
# Extra tidying:
norway_tidy2 <- tidy_data2(norway_tidy)
}

# See ?run_assessment_tar for full workflow

}
